requirements.txt
# EO Evaluation Framework Dependencies
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.0
tqdm>=4.65.0
google-genai>=1.0.0
anthropic>=0.40.0
streamlit>=1.30.0



